requirements.txt
certifi==2023.5.7
mysqlclient==2.2.0
PyMySQL==1.1.0
python-dotenv==1.0.0
